Updated for app store release. Added iPad 12.9" screenshot. Changed support email address back to me.

+++ externals/CocoaHTTPServer/HTTPConnection.m
@@ -954,7 +954,7 @@ static NSMutableArray *recentNonces;
 **/
 - (NSData *)chunkedTransferSizeLineForLength:(NSUInteger)length
 {
-	return [[NSString stringWithFormat:@"%lx\r\n", length] dataUsingEncoding:NSUTF8StringEncoding];
+	return [[NSString stringWithFormat:@"%lx\r\n", (unsigned long)length] dataUsingEncoding:NSUTF8StringEncoding];
 }
 
 /**
